backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Telegram Libraries
aiogram==3.4.1
telethon==1.34.0
cryptg==0.4.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Task Queue
celery==5.3.6
redis==5.0.1

# OpenAI
openai==1.10.0

# Scheduling
apscheduler==3.10.4

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
pytz==2024.1
openpyxl==3.1.2
qrcode==7.4.2
pillow==10.2.0
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Web Scraping
beautifulsoup4==4.12.3
